refactor(bridge): extract download file ID codec
